django/pyproject.toml
[project]
name = "docurba"
version = "1"
requires-python = ">=3.13"

dependencies = [
    "django",
    "django-browser-reload",
    "django-debug-toolbar",
    "django-environ",
    "django-extensions",
    "django-revproxy",
    "djangorestframework",
    "gunicorn[gevent]",
    "psycopg[binary,pool]",
    "sentry-sdk[django]",
    "whitenoise[brotli]",
]

[dependency-groups]
dev = [
    "honcho",
    "polars",
    "ptpython",
    "pytest-cov",
    "pytest-django",
    "ruff",
    "pre-commit>=4.3.0", # https://github.com/pre-commit/pre-commit
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
addopts = "--reuse-db --cov=."

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812", # Missing trailing comma, conflicts with formatter
    "D10",    # Missing docstring
    "E501",   # Line too long
    "TD",     # Todos
    # "ERA001", # Commented out code
]
[tool.ruff.lint.per-file-ignores]
"*test_*.py" = [
    "S101",    # asserts allowed in tests...
    "PLR2004", # Magic value used in comparison
]
[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings.
convention = "pep257"
